+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Command-line option `--check` to check if changelog has any changes under \[Unreleased\].
 
 ## [1.1.0] - 2018-01-02
 ### Added